requirements.txt
alembic
bcrypt
cffi
click
coverage
Flask
Flask-Bcrypt
Flask-Cache
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask-Testing
itsdangerous
Jinja2
Mako
MarkupSafe
psycopg2-binary
pycparser
PyJWT
python-editor
six
SQLAlchemy
Werkzeug
